Let the PTY thread worry about its main function
